# docs/tugas-5/openapi-gateway.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/destinasi": {
      "get": {
        "operationId": "DestinasiV1Controller_findAll[0]_v1",
        "parameters": [
          {
            "name": "kategori",
            "required": false,
            "in": "query",
            "description": "Filter berdasarkan nama kategori",
            "schema": {
              "example": "Pantai",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Nomor halaman, dipakai bersama limit",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Jumlah data per halaman. Tanpa limit, seluruh data dikembalikan.",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Destinasi"
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "service-destinasi tidak dapat dihubungi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Menampilkan daftar destinasi, dapat difilter berdasarkan kategori",
        "tags": [
          "Destinasi v1"
        ]
      },
      "post": {
        "operationId": "DestinasiV1Controller_create[0]_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDestinasiDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Destinasi"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "service-destinasi tidak dapat dihubungi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Menambahkan destinasi baru (khusus admin)",
        "tags": [
          "Destinasi v1"
        ]
      }
    },
    "/destinasi": {
      "get": {
        "operationId": "DestinasiV1Controller_findAll[1]",
        "parameters": [
          {
            "name": "kategori",
            "required": false,
            "in": "query",
            "description": "Filter berdasarkan nama kategori",
            "schema": {
              "example": "Pantai",
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Nomor halaman, dipakai bersama limit",
            "schema": {
              "minimum": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Jumlah data per halaman. Tanpa limit, seluruh data dikembalikan.",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Destinasi"
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "service-destinasi tidak dapat dihubungi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Menampilkan daftar destinasi, dapat difilter berdasarkan kategori",
        "tags": [
          "Destinasi v1"
        ]
      },
      "post": {
        "operationId": "DestinasiV1Controller_create[1]",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDestinasiDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Destinasi"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "service-destinasi tidak dapat dihubungi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Menambahkan destinasi baru (khusus admin)",
        "tags": [
          "Destinasi v1"
        ]
      }
    },
    "/v1/destinasi/{id}": {
      "get": {
        "operationId": "DestinasiV1Controller_findOne[0]_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Destinasi"
                }
              }
            }
          },
          "400": {
            "description": "id bukan angka",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "service-destinasi tidak dapat dihubungi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Detail satu destinasi",
        "tags": [
          "Destinasi v1"
        ]
      },
      "patch": {
        "operationId": "DestinasiV1Controller_update[0]_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDestinasiDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Destinasi"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "service-destinasi tidak dapat dihubungi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mengubah sebagian data destinasi (khusus admin)",
        "tags": [
          "Destinasi v1"
        ]
      },
      "delete": {
        "operationId": "DestinasiV1Controller_remove[0]_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Destinasi berhasil dihapus"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Destinasi sudah memiliki reservasi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "service-destinasi tidak dapat dihubungi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Menghapus destinasi (khusus admin)",
        "tags": [
          "Destinasi v1"
        ]
      }
    },
    "/destinasi/{id}": {
      "get": {
        "operationId": "DestinasiV1Controller_findOne[1]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Destinasi"
                }
              }
            }
          },
          "400": {
            "description": "id bukan angka",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "service-destinasi tidak dapat dihubungi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Detail satu destinasi",
        "tags": [
          "Destinasi v1"
        ]
      },
      "patch": {
        "operationId": "DestinasiV1Controller_update[1]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDestinasiDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Destinasi"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "service-destinasi tidak dapat dihubungi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mengubah sebagian data destinasi (khusus admin)",
        "tags": [
          "Destinasi v1"
        ]
      },
      "delete": {
        "operationId": "DestinasiV1Controller_remove[1]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "message": "Destinasi berhasil dihapus"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Destinasi sudah memiliki reservasi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "service-destinasi tidak dapat dihubungi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Menghapus destinasi (khusus admin)",
        "tags": [
          "Destinasi v1"
        ]
      }
    },
...
    "/v1/destinasi/{id}/ulasan": {
      "get": {
        "operationId": "DestinasiV1Controller_findUlasan[0]_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Ulasan"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "service-destinasi tidak dapat dihubungi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Ulasan milik destinasi tertentu, terbaru lebih dulu",
        "tags": [
          "Destinasi v1"
        ]
      }
    },
    "/destinasi/{id}/ulasan": {
      "get": {
        "operationId": "DestinasiV1Controller_findUlasan[1]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Ulasan"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "service-destinasi tidak dapat dihubungi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Ulasan milik destinasi tertentu, terbaru lebih dulu",
        "tags": [
          "Destinasi v1"
        ]
      }
    },
    "/v1/destinasi/{id}/fasilitas": {
      "get": {
        "operationId": "DestinasiV1Controller_findFasilitas[0]_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Fasilitas"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "service-destinasi tidak dapat dihubungi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Fasilitas yang tersedia di destinasi tertentu",
        "tags": [
          "Destinasi v1"
        ]
      }
    },
    "/destinasi/{id}/fasilitas": {
      "get": {
        "operationId": "DestinasiV1Controller_findFasilitas[1]",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Fasilitas"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "service-destinasi tidak dapat dihubungi",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Fasilitas yang tersedia di destinasi tertentu",
        "tags": [
          "Destinasi v1"
        ]
      }
    },
...
  },
  "info": {
    "title": "WisataKu API Gateway",
    "description": "Titik masuk tunggal WisataKu. Meneruskan request ke service-destinasi dan service-reservasi melalui TCP. Mendukung URI versioning: /v1 (hargaTiket) dan /v2 (hargaDewasa, hargaAnak); path tanpa prefix versi setara v1.",
    "version": "2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404
          },
          "message": {
            "oneOf": [
              {
                "type": "string"
              },
              {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            ],
            "example": "Destinasi dengan id 999 tidak ditemukan"
          },
          "error": {
            "type": "string",
            "example": "Not Found"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
...
      "Destinasi": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "nama": {
            "type": "string",
            "example": "Pantai Kuta Mandalika"
          },
          "kategori": {
            "type": "string",
            "example": "Pantai"
          },
          "lokasi": {
            "type": "string",
            "example": "Kuta, Lombok Tengah",
            "nullable": true
          },
          "deskripsi": {
            "type": "string",
            "example": "Pantai berpasir putih di kawasan Mandalika.",
            "nullable": true
          },
          "hargaTiket": {
            "type": "number",
            "example": 15000
          },
          "ratingRata": {
            "type": "number",
            "example": 4.5
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-29T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "nama",
          "kategori",
          "lokasi",
          "deskripsi",
          "hargaTiket",
          "ratingRata",
          "createdAt"
        ]
      },
      "Ulasan": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "destinasiId": {
            "type": "number",
            "example": 1
          },
          "userId": {
            "type": "number",
            "example": 2
          },
          "rating": {
            "type": "number",
            "example": 5,
            "minimum": 1,
            "maximum": 5
          },
          "komentar": {
            "type": "string",
            "example": "Pemandangan luar biasa!"
          },
          "tanggal": {
            "format": "date-time",
            "type": "string",
            "example": "2026-09-29T10:00:00.000Z"
          }
        },
        "required": [
          "id",
          "destinasiId",
          "userId",
          "rating",
          "komentar",
          "tanggal"
        ]
      },
...
      "CreateDestinasiDto": {
        "type": "object",
        "properties": {
          "nama": {
            "type": "string",
            "example": "Pantai Kuta Mandalika",
            "maxLength": 150
          },
          "kategori": {
            "type": "string",
            "example": "Pantai",
            "description": "Nama kategori; dibuat otomatis jika belum ada"
          },
          "lokasi": {
            "type": "string",
            "example": "Kuta, Lombok Tengah",
            "maxLength": 150
          },
          "deskripsi": {
            "type": "string",
            "example": "Pantai berpasir putih di kawasan Mandalika."
          },
          "hargaTiket": {
            "type": "number",
            "example": 15000,
            "minimum": 0
          }
        },
        "required": [
          "nama",
          "kategori",
          "hargaTiket"
        ]
      },
      "UpdateDestinasiDto": {
        "type": "object",
        "properties": {
          "nama": {
            "type": "string",
            "example": "Pantai Kuta Mandalika",
            "maxLength": 150
          },
          "kategori": {
            "type": "string",
            "example": "Pantai",
            "description": "Nama kategori; dibuat otomatis jika belum ada"
          },
          "lokasi": {
            "type": "string",
            "example": "Kuta, Lombok Tengah",
            "maxLength": 150
          },
          "deskripsi": {
            "type": "string",
            "example": "Pantai berpasir putih di kawasan Mandalika."
          },
          "hargaTiket": {
            "type": "number",
            "example": 15000,
            "minimum": 0
          }
        }
      },
...
    }
  }
}